Trimming innerHTML and expected output
This test is VERY strict about spacing. While I think indentation of the elements is important, the spacing before the h2 and after the ul is arbitrary. Unless we are teaching some sort of best practice for using string templates, or if we really want to force fellows to look at the tests (in which case they'd just see the answer anyway), there's no need for this test to be as strict as it is

A better test would check the structure of the `bioEl` Node by Node but I'm guessing that would be more than what you want to do.